# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API para gestão integral da produção gráfica.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals/{token}:
    get:
      operationId: getSampleApproval
      tags: [approvals]
      summary: Consultar amostra através de link seguro
      parameters:
        - { name: token, in: path, required: true, schema: { type: string, minLength: 32 } }
      responses:
        "200":
          description: Amostra pendente ou decisão já registada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalView" }
        "404": { $ref: "#/components/responses/NotFound" }
  /approvals/{token}/file:
    get:
      operationId: getSampleApprovalFile
      tags: [approvals]
      summary: Abrir o ficheiro da amostra através do link seguro
      parameters:
        - { name: token, in: path, required: true, schema: { type: string, minLength: 32 } }
      responses:
        "200":
          description: Ficheiro da amostra.
          content:
            application/pdf:
              schema: { type: string, format: binary }
            image/*:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/NotFound" }
  /approvals/{token}/decision:
    post:
      operationId: decideSampleApproval
      tags: [approvals]
      summary: Aprovar ou rejeitar a amostra
      parameters:
        - { name: token, in: path, required: true, schema: { type: string, minLength: 32 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SampleDecisionInput" }
      responses:
        "200":
          description: Decisão registada no histórico.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalView" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    BadRequest:
      description: Dados inválidos ou transição não permitida.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: Registo não encontrado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ApprovalView:
      type: object
      required: [ticketCode, clientName, company, product, quantity, dueDate, version, fileName, contentType, status, submittedAt, decisionComment]
      properties:
        ticketCode: { type: string }
        clientName: { type: string }
        company: { type: [string, "null"] }
        product: { type: string }
        quantity: { type: integer }
        dueDate: { type: string, format: date }
        version: { type: integer }
        fileName: { type: string }
        contentType: { type: string }
        status: { type: string }
        submittedAt: { type: string, format: date-time }
        decisionComment: { type: [string, "null"] }
    SampleDecisionInput:
      type: object
      required: [decision, approverName]
      properties:
        decision: { type: string, enum: [APROVAR, REJEITAR] }
        approverName: { type: string, minLength: 1 }
        comment: { type: [string, "null"] }
...