# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API para gestão integral da produção gráfica.
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}